Handle the case when the info.xml file is not created (e.g., DESeq, possibly EdgeR).
git-svn-id: https://pbtech-vc.med.cornell.edu/public/svn/icb/trunk/gobyweb/server-conf/plugins@32967 60cd2ccc-2b05-0410-9566-da273a940bff

+++ analyses/DIFF_EXP_GOBY/parallel.sh
@@ -133,12 +133,12 @@ function run_fdr() {
    # The following sections extracts the info.xml file stored among split-results
    # and adjust the PART_RESULT_FILES variables to exclude the fake tsv info file.
    INFO_FILE=`ls -1 ${PART_RESULT_FILES} |grep info`
-   cp ${INFO_FILE} ./info.xml
+   if [ ${INFO_FILE} != "" ] && [ -e ${INFO_FILE} ]; then
+       cp ${INFO_FILE} ./info.xml
+       # Run FDR to combine parts:
 
-   # Run FDR to combine parts:
-
-   PART_RESULT_FILES=`echo ${PART_RESULT_FILES} | sed -e 's!'${INFO_FILE}'!!'`
-
+       PART_RESULT_FILES=`echo ${PART_RESULT_FILES} | sed -e 's!'${INFO_FILE}'!!'`
+   fi
    OUT_FILENAME=combined-stats.tsv
    run-goby ${PLUGIN_NEED_COMBINE_JVM} fdr \
           --column-selection-filter t-test  \
